refactor(frontend): split chatterbox voice modal into section components
750-line modal becomes a 240-line composition over three verbatim-moved
sections (runtime overview, voice creator, preview player); non-view
helpers join the existing chatterbox-voice-model.ts per the model/view
split.
